outsource algorithms into a proper hierachy
