+++ optim_algos.py
@@ -1,3 +1,4 @@
+from abc import ABC, abstractmethod
 from typing import Union
 
 import numpy as np
@@ -5,7 +6,8 @@ from numpy.linalg import LinAlgError
 
 import func
 
-class LineSearch:
+
+class LineSearch(ABC):
     def __init__(self, f: func.Function,
                  name: str,
                  start_point: np.ndarray = None,
@@ -54,6 +56,7 @@ class LineSearch:
         print(f"Final residual norm is {self.residual_norm()}")
         print('-' * 50)
 
+    @abstractmethod
     def update(self):
         self.iterations += 1
         # optional printing for seeing the process
@@ -76,124 +79,6 @@ class LineSearch:
 
     def residual_norm(self):
         return np.linalg.norm(self.grad_f_k, ord=self.norm)
-
-
-class NewtonFamilyMethod(LineSearch):
-    def __init__(self, f: func.Function,
-                 name: str = 'SD',
-                 start_point: np.ndarray = None,
-                 norm: Union[str, float] = 2,
-                 eps: float = 10**-6,
-                 max_iterations: int = 10 ** 6,
-                 initial_alpha: float = 1,
-                 rho: float = 0.99,
-                 c: float = 0.99):
-        # SD - Steepest Descent
-        # NM - Newton Method
-        # QN - Quasi-Newton Method
-        if name not in ['SD', 'QN', 'NM', 'QN-BFGS']: #added bfgs - other way???
-            raise ValueError("Possible names for NewtonFamilyMethod are: 'SD', 'NM' and 'QN'")
-        super().__init__(f, name, start_point, norm, eps, max_iterations, initial_alpha, rho, c)
-        self.H = np.eye(self.f.get_dim())  # inital approx. of inverse Hessian - added ???
-
-    def compute_B(self):
-        if self.name == 'SD':
-                B_k = np.diag(np.ones(shape=self.f.get_dim()))
-
-        elif self.name == 'NM':
-            if np.allclose(self.f.num_hessian(self.x_k), np.zeros(shape=self.f.get_dim())):
-                if np.allclose(self.f.hessian(self.x_k), np.zeros(shape=self.f.get_dim())):
-                    print("DROPPED TO SD")
-                    B_k = np.diag(np.ones(shape=self.f.get_dim()))
-                else:
-                    print("NUM ISSUES WITH HESSIAN")
-                    if self.f.get_dim() > 1:
-                        B_k = np.linalg.inv(self.f.hessian(self.x_k))
-                    else:
-                        B_k = np.array([1 / self.f.hessian(self.x_k)])
-            else:
-                try:
-                    if self.f.get_dim() > 1:
-                        B_k = np.linalg.inv(self.f.num_hessian(self.x_k))
-                    else:
-                        B_k = np.array([1 / self.f.num_hessian(self.x_k)])
-                except LinAlgError:
-                    print("DROPPED TO SD")
-                    B_k = np.diag(np.ones(shape=self.f.get_dim()))
-
-            """if self.f.get_dim() > 1:
-                try:
-                    B_k = np.linalg.inv(self.f.num_hessian(self.x_k))
-                except LinAlgError:
-                    print(f"Problems with inverting numerical hessian at point x_k = {self.x_k},"
-                          f" which is {self.f.num_hessian(self.x_k)}")
-                    try:
-                        B_k = np.linalg.inv(self.f.hessian(self.x_k))
-                    except LinAlgError:
-                        print(f"Explicit hessian failed at point x_k = {self.x_k},"
-                              f" which is {self.f.hessian(self.x_k)}")
-                        raise LinAlgError"""
-            """else:
-                try:
-                    B_k = np.array([1 / self.f.num_hessian(self.x_k)])
-                except ZeroDivisionError:
-                    print(f"Problems with inverting numerical hessian at point x_k = {self.x_k},"
-                          f" which is {self.f.num_hessian(self.x_k)}")
-                    try:
-                        B_k = np.array([1 / self.f.hessian(self.x_k)])
-                    except ZeroDivisionError:
-                        print(f"Explicit hessian failed at point x_k = {self.x_k},"
-                              f" which is {self.f.hessian(self.x_k)}")
-                        raise LinAlgError"""
-        elif 'QN' in self.name:
-            I = np.eye(self.f.get_dim())
-            x_k = self.x_k
-            grad_f_k = self.f.grad(x_k)
-            self.p_k = -self.H @ grad_f_k  # search direction (6.18)
-
-            alpha_k = LineSearch.compute_alpha_k(
-                self)  # backtracking line search # same for SR1 ?
-
-            x_new = x_k + alpha_k * self.p_k  # (6.3)
-            # define s_k and y_k (6.5)
-            s_k = x_new - x_k  # or: alpha_k * p_k #for sk1?
-            grad_f_new = self.f.grad(x_new)
-            y_k = grad_f_new - grad_f_k
-
-            if self.name == 'QN-BFGS':
-                if np.abs(y_k @ s_k) >= 0:  # (6.7) check curvature condition?
-                    # (6.17) compute H_{k+1} -> H_new using BFGS formula
-                    rho_k = 1.0 / (y_k.T @ s_k)  # (6.14)
-                    self.H = (I - rho_k * s_k @ y_k.T) @ self.H @ (I - rho_k * y_k @ s_k.T) + rho_k * s_k @ s_k.T
-            else: # SR1
-                if np.abs(np.dot(y_k, s_k)) >= self.eps * np.linalg.norm(y_k) * np.linalg.norm(s_k):
-                    Hy = np.dot(self.H, y_k)
-                    self.H += np.outer(s_k - Hy, s_k - Hy) / np.dot(y_k, s_k - Hy)
-            self.x_k = x_new
-            self.grad_f_k = grad_f_new
-            B_k = self.H
-        else:
-            # decide between 'SK1' and 'BFGS' or just do both???
-            raise NotImplementedError("Quasi Newton Method is not implemented yet")
-
-        return B_k
-
-    def compute_p_k(self):
-        return - self.compute_B() @ self.grad_f_k
-
-    def update(self):
-        if self.iterations == 0 or self.alpha_k == 1 or (self.grad_f_k != self.f.grad(self.x_k)).any():
-            # when next grad is the same as current, but alpha isn't 1, we are stuck
-            # it could happen that the optimal step size is more than -1*grad, but than alpha should be 1
-            self.grad_f_k = self.f.grad(self.x_k)
-            self.p_k = self.compute_p_k()
-            self.alpha_k = self.compute_alpha_k()
-
-            self.x_k += self.alpha_k * self.p_k
-            super().update()
-        else:
-            self.stuck = True
-
 
 class ConjugateGradient(LineSearch):
     def __init__(self, f: func.Function,
